requirements.txt
asgiref==3.3.4
astroid==2.5.6
Django==3.2.4
djangorestframework==3.12.4
djangorestframework-camel-case==1.2.0
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
#pkg-resources==0.0.0
psycopg2-binary==2.8.6
PyJWT==2.1.0
pylint==2.8.3
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
wrapt==1.12.1
